requirements.txt
biplist
pybonjour
pycrypto
pygame
pysdl2
